requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
dateparser==1.2.0
Django==4.2.10
django-bootstrap4==24.1
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-stdimage==6.0.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
kombu==5.3.5
Library==0.0.0
moment==0.12.1
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.2
regex==2023.12.25
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
times==0.7
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
